Update SDK to v1.5.7

+++ Assets/NRSDK/Scripts/Capture/NRVideoCapture.cs
@@ -12,31 +12,31 @@ namespace NRKernal.Record
     using UnityEngine;
     using NRKernal;
     using System;
-    using System.IO;
     using System.Collections.Generic;
 
     /// <summary>
-    /// Records a video from the MR images directly to disk.
-    /// MR images comes from rgb camera or rgb camera image and virtual image blending.
-    /// The final video recording will be stored on the file system in the MP4 format.
-    /// </summary>
+    /// Records a video from the MR images directly to disk. MR images comes from rgb camera or rgb
+    /// camera image and virtual image blending. The final video recording will be stored on the file
+    /// system in the MP4 format. </summary>
     public class NRVideoCapture : IDisposable
     {
+        /// <summary> Default constructor. </summary>
         public NRVideoCapture()
         {
             IsRecording = false;
         }
 
+        /// <summary> Finalizer. </summary>
         ~NRVideoCapture()
         {
 
         }
 
+        /// <summary> The supported resolutions. </summary>
         private static IEnumerable<Resolution> m_SupportedResolutions = null;
 
-        /// <summary>
-        /// A list of all the supported device resolutions for recording videos.
-        /// </summary>
+        /// <summary> A list of all the supported device resolutions for recording videos. </summary>
+        /// <value> The supported resolutions. </value>
         public static IEnumerable<Resolution> SupportedResolutions
         {
             get
@@ -46,7 +46,7 @@ namespace NRKernal.Record
                     var resolutions = new List<Resolution>();
                     var resolution = new Resolution();
 #if !UNITY_EDITOR
-                    NativeResolution rgbResolution = NRDevice.Instance.NativeHMD.GetEyeResolution(NativeEye.RGB);
+                    NativeResolution rgbResolution = NRDevice.Instance.NativeHMD.GetEyeResolution((int)NativeEye.RGB);
 #else
                     NativeResolution rgbResolution = new NativeResolution(1280, 720);
 #endif
@@ -61,17 +61,22 @@ namespace NRKernal.Record
         }
 
         /// <summary>
-        /// Indicates whether or not the VideoCapture instance is currently recording video.
-        /// </summary>
+        /// Indicates whether or not the VideoCapture instance is currently recording video. </summary>
+        /// <value> True if this object is recording, false if not. </value>
         public bool IsRecording { get; private set; }
 
+        /// <summary> Context for the capture. </summary>
         private FrameCaptureContext m_CaptureContext;
 
+        /// <summary> Gets the context. </summary>
+        /// <returns> The context. </returns>
         public FrameCaptureContext GetContext()
         {
             return m_CaptureContext;
         }
 
+        /// <summary> Gets the preview texture. </summary>
+        /// <value> The preview texture. </value>
         public Texture PreviewTexture
         {
             get
@@ -80,6 +85,9 @@ namespace NRKernal.Record
             }
         }
 
+        /// <summary> Creates an asynchronous. </summary>
+        /// <param name="showHolograms">     True to show, false to hide the holograms.</param>
+        /// <param name="onCreatedCallback"> The on created callback.</param>
         public static void CreateAsync(bool showHolograms, OnVideoCaptureResourceCreatedCallback onCreatedCallback)
         {
             NRVideoCapture capture = new NRVideoCapture();
@@ -88,10 +96,9 @@ namespace NRKernal.Record
         }
 
         /// <summary>
-        /// Returns the supported frame rates at which a video can be recorded given a resolution.
-        /// </summary>
-        /// <param name="resolution">A recording resolution.</param>
-        /// <returns>The frame rates at which the video can be recorded.</returns>
+        /// Returns the supported frame rates at which a video can be recorded given a resolution. </summary>
+        /// <param name="resolution"> A recording resolution.</param>
+        /// <returns> The frame rates at which the video can be recorded. </returns>
         public static IEnumerable<int> GetSupportedFrameRatesForResolution(Resolution resolution)
         {
             List<int> frameRates = new List<int>();
@@ -99,9 +106,7 @@ namespace NRKernal.Record
             return frameRates;
         }
 
-        /// <summary>
-        /// Dispose must be called to shutdown the PhotoCapture instance.
-        /// </summary>
+        /// <summary> Dispose must be called to shutdown the PhotoCapture instance. </summary>
         public void Dispose()
         {
             if (m_CaptureContext != null)
@@ -111,6 +116,9 @@ namespace NRKernal.Record
             }
         }
 
+        /// <summary> Starts recording asynchronous. </summary>
+        /// <param name="filename">                        Filename of the file.</param>
+        /// <param name="onStartedRecordingVideoCallback"> The on started recording video callback.</param>
         public void StartRecordingAsync(string filename, OnStartedRecordingVideoCallback onStartedRecordingVideoCallback)
         {
             var result = new VideoCaptureResult();
@@ -139,6 +147,9 @@ namespace NRKernal.Record
             }
         }
 
+        /// <summary> Starts video mode asynchronous. </summary>
+        /// <param name="setupParams">                Options for controlling the setup.</param>
+        /// <param name="onVideoModeStartedCallback"> The on video mode started callback.</param>
         public void StartVideoModeAsync(CameraParameters setupParams, OnVideoModeStartedCallback onVideoModeStartedCallback)
         {
             setupParams.camMode = CamMode.VideoMode;
@@ -149,6 +160,8 @@ namespace NRKernal.Record
             onVideoModeStartedCallback?.Invoke(result);
         }
 
+        /// <summary> Stops recording asynchronous. </summary>
+        /// <param name="onStoppedRecordingVideoCallback"> The on stopped recording video callback.</param>
         public void StopRecordingAsync(OnStoppedRecordingVideoCallback onStoppedRecordingVideoCallback)
         {
             var result = new VideoCaptureResult();
@@ -175,6 +188,8 @@ namespace NRKernal.Record
             }
         }
 
+        /// <summary> Stops video mode asynchronous. </summary>
+        /// <param name="onVideoModeStoppedCallback"> The on video mode stopped callback.</param>
         public void StopVideoModeAsync(OnVideoModeStoppedCallback onVideoModeStoppedCallback)
         {
             m_CaptureContext.StopCaptureMode();
@@ -183,9 +198,7 @@ namespace NRKernal.Record
             onVideoModeStoppedCallback?.Invoke(result);
         }
 
-        /// <summary>
-        /// Contains the result of the capture request.
-        /// </summary>
+        /// <summary> Contains the result of the capture request. </summary>
         public enum CaptureResultType
         {
             /// <summary>
@@ -226,23 +239,18 @@ namespace NRKernal.Record
         //}
 
         /// <summary>
-        ///  A data container that contains the result information of a video recording operation.
-        /// </summary>
+        /// A data container that contains the result information of a video recording operation. </summary>
         public struct VideoCaptureResult
         {
             /// <summary>
-            /// A generic result that indicates whether or not the VideoCapture operation succeeded.
-            /// </summary>
+            /// A generic result that indicates whether or not the VideoCapture operation succeeded. </summary>
             public CaptureResultType resultType;
 
-            /// <summary>
-            /// The specific HResult value.
-            /// </summary>
+            /// <summary> The specific HResult value. </summary>
             public long hResult;
 
-            /// <summary>
-            /// Indicates whether or not the operation was successful.
-            /// </summary>
+            /// <summary> Indicates whether or not the operation was successful. </summary>
+            /// <value> True if success, false if not. </value>
             public bool success
             {
                 get
@@ -252,34 +260,25 @@ namespace NRKernal.Record
             }
         }
 
-        /// <summary>
-        /// Called when the web camera begins recording the video.
-        /// </summary>
-        /// <param name="result">Indicates whether or not video recording started successfully.</param>
+        /// <summary> Called when the web camera begins recording the video. </summary>
+        /// <param name="result"> Indicates whether or not video recording started successfully.</param>
         public delegate void OnStartedRecordingVideoCallback(VideoCaptureResult result);
 
-        /// <summary>
-        ///  Called when a VideoCapture resource has been created.
-        /// </summary>
-        /// <param name="captureObject">The VideoCapture instance.</param>
+        /// <summary> Called when a VideoCapture resource has been created. </summary>
+        /// <param name="captureObject"> The VideoCapture instance.</param>
         public delegate void OnVideoCaptureResourceCreatedCallback(NRVideoCapture captureObject);
 
-        /// <summary>
-        /// Called when video mode has been started.
-        /// </summary>
-        /// <param name="result">Indicates whether or not video mode was successfully activated.</param>
+        /// <summary> Called when video mode has been started. </summary>
+        /// <param name="result"> Indicates whether or not video mode was successfully activated.</param>
         public delegate void OnVideoModeStartedCallback(VideoCaptureResult result);
 
-        /// <summary>
-        /// Called when video mode has been stopped.
-        /// </summary>
-        /// <param name="result">Indicates whether or not video mode was successfully deactivated.</param>
+        /// <summary> Called when video mode has been stopped. </summary>
+        /// <param name="result"> Indicates whether or not video mode was successfully deactivated.</param>
         public delegate void OnVideoModeStoppedCallback(VideoCaptureResult result);
 
-        /// <summary>
-        ///  Called when the video recording has been saved to the file system.
-        /// </summary>
-        /// <param name="result">Indicates whether or not video recording was saved successfully to the file system.</param>
+        /// <summary> Called when the video recording has been saved to the file system. </summary>
+        /// <param name="result"> Indicates whether or not video recording was saved successfully to the
+        ///                       file system.</param>
         public delegate void OnStoppedRecordingVideoCallback(VideoCaptureResult result);
     }
 }
